requirements.txt
torch
fair-esm
numpy